Add files via upload
Quiz management system:
In this project, the user can read the notes and select the subject from the quiz management system. after successfully selecting the user can answer the MCQs with help of the radio button and see it's a score at the end of the quiz when user can press the finish button.